Fixing conflicts
