+++ client/starter-app/.angular-cli.json
@@ -22,8 +22,6 @@
             "../node_modules/bootstrap/dist/css/bootstrap.min.css"
         ],
         "scripts": [
-            "../node_modules/bootstrap/dist/js/bootstrap.min.js"
-
         ],
         "environmentSource": "environments/environment.ts",
         "environments": {
