+++ client/starter-app/src/app/app-routing.module.ts
@@ -3,7 +3,6 @@ import { Routes, RouterModule } from '@angular/router';
 import { ProjectComponent } from './pages/project/project.component';
 import { LoginPageComponent } from './pages/login-page/login-page.component';
 import { WeeklyTsComponent } from './pages/weekly-ts/weekly-ts.component';
-import { FormPageComponent } from './pages/form-page/form-page.component';
 
 const routes: Routes = [
   {
