+++ client/starter-app/src/app/providers/weekly-ts.service.ts
@@ -1,6 +1,7 @@
 import { Injectable } from '@angular/core';
 import {  Http,Response,Headers} from '@angular/http';
 import 'rxjs/add/operator/map';
+import 'rxjs/add/operator/toPromise';
 import * as moment from 'moment';
 
 @Injectable()
@@ -10,15 +11,10 @@ export class WeeklyTsService {
 
   private apiURL = 'http://10.0.1.30:8080/api/'
 
-  postDate(val){
-      let postObject = {
-        'startDate': '',
-        'toDate': ''
-      };
-      postObject.startDate = moment().month(val.month - 1).date(val.range[0]).format('YYYY-MM-DD');
-      postObject.toDate = moment().month(val.month - 1).date(val.range[1]).format('YYYY-MM-DD');
-      return this.http.post(this.apiURL + 'employees/timesheet', postObject)
-        .map((response:Response) => response.json());
+    postDate(val){
+      return this.http.post(this.apiURL + 'employees/timesheetRetrieval', val)
+        .toPromise()
+        .then(res => console.log(res));
     }
 
 }
